requirements/dev.txt
django-debug-toolbar
django-extensions